config.py
from decouple import config

BOT_TOKEN = config("BOT_TOKEN", default="")

APP_ICON = config("APP_ICON", default="https://f004.backblazeb2.com/file/switch-bucket/e78c287a-abc0-11ee-883b-d41b81d4a9ef.png")
SECONDARY_ICON = config("SECONDARY_ICON", default="https://f004.backblazeb2.com/file/switch-bucket/6c0928da-abc3-11ee-82f3-d41b81d4a9ef.png")
