requirements.txt
aiohttp
requests
swibots
python-decouple
apscheduler
beautifulsoup4